add fontawesom and add first comment

+++ src/main/java/com/sofency/community/mapper/QuestionCustomMapper.java
@@ -9,6 +9,6 @@ import com.sofency.community.pojo.Question;
  */
 
 public interface QuestionCustomMapper {
-    public void incrView(Integer id);
+    public void incrView(Long id);
     public void incrCommentCount(Question question);
 }
